Refactor: Separate types and utilities into failed-calls-types.ts
